requirements.txt
altair==5.5.0
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.7.0
attrs==25.1.0
beautifulsoup4==4.12.3
blinker==1.9.0
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
distro==1.9.0
et_xmlfile==2.0.0
extruct==0.18.0
fastapi==0.116.2
Flask==3.1.1
flask-cors==6.0.1
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.181.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-generativeai==0.8.5
google_search_results==2.4.2
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpcio==1.75.0
grpcio-status==1.71.2
gTTS==2.5.4
h11==0.14.0
html5lib==1.1
html_text==0.7.0
httpcore==1.0.7
httplib2==0.30.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jstyleson==0.0.2
keyboard==0.13.5
lxml==5.3.0
lxml_html_clean==0.4.2
markdownify==0.14.1
MarkupSafe==3.0.2
mcp==1.1.2
mcp-server-fetch==0.6.2
mf2py==2.0.1
narwhals==2.5.0
numpy==2.3.2
oauthlib==3.3.1
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.2
pandas==2.3.1
pillow==11.3.0
platformdirs==4.3.6
playwright==1.55.0
Protego==0.3.1
proto-plus==1.26.1
protobuf==5.29.5
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
pydeck==0.9.1
pyee==13.0.0
pyparsing==3.2.3
pyRdfa3==3.6.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
rdflib==7.1.4
readabilipy==0.3.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rpds-py==0.27.1
rsa==4.9.1
schedule==1.2.2
selenium==4.33.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sse-starlette==2.2.1
starlette==0.43.0
streamlit==1.49.1
tenacity==9.1.2
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.4.0
uvicorn==0.35.0
w3lib==2.3.1
watchdog==6.0.0
watchfiles==1.1.0
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wsproto==1.2.0
xlsxwriter==3.2.5
zeep==4.3.1
